Add catch-all for exception when parsing battery cycles
Some devices do not have a `cycle_count` file, which leads to an
error when trying to parse the cycle count value. Unfortunately,
this is different from the out-of-range exception caught by the
current implementation, and as a result the program crashes with
a SIGABRT (tested on macOS). To resolve this, add a catch-all for
the cycle count since it is not terribly important to calculate
the battery health anyway.